pyproject.toml
[project]
name = "memex-md-mcp"
version = "0.1.0"
description = "MCP server for semantic search over markdown vaults"
readme = "README.md"
requires-python = ">=3.14"
license = "MIT"
authors = [{ name = "Maximilian Wolf" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "mcp[cli]>=1.24.0",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.0",
]

[project.scripts]
memex-md-mcp = "memex_md_mcp:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
    "ty>=0.0.1a1",
]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM108", # ternary operator - often less readable
]

[tool.ruff.lint.isort]
known-first-party = ["memex_md_mcp"]

[tool.ty.environment]
python-version = "3.14"
